Moving forward to vmap perspectives, I have started writing a Perspective submodule for DNA.res that will permute DNAs to match a given interpretation order. The old formform always re-calculates a FORM to obtain a new formDNA for each given variable order, but this is very inefficient considering they are just permutations on the DNA structure. However, I found my first naive approach fails on tests with DNA of size 3 where the permutation from the base DNA involves swapping more than 1 dimension. I’ll have to gain a deeper understanding of the symmetries involved before making another attempt. Certainly there is also a more standard solution to this problem, I just have to find it.
I also started creating VMapGeom.res, which will be an extension module to generate the actual geometric parameters for VMap representations. Right now I am unsure how much data I need to attach to the structure and if it is wise to reproduce the tree from the core VMap structure when it could actually be generated in one go (as I did in formform v1).

In my attempt to convert NMUI-coded DNA number strings (from my old FORM tricorder) to the new NUIM ordering, I found that the conversion methods in DNA.res should also reorder the DNA to NUIM format since otherwise the ordering will still match NMUI, which is not expected from my internal implementation of formDNA.

Another major change concerns the FCON and FVAR modules, which are now submodules of FORM and renamed to “ConstFORM” and “VarFORM” to be more descriptive. I was thinking about getting rid of them altogether, but ended up just removing their type aliases and constrained them to methods that directly manipulate these specific structures, but moved methods like getVar to FORM so the API becomes simpler.

Finally, I renamed VTable in Value.res to VDict, as it is a more accurate description of the data structure. As with VMap, I decided to have the make function depend on VSpace instead of DNA, which would already be too specific.