There have been a lot of changes since my last commit and I don’t remember too much of it, since unfortunately I have postponed this commit for many months because of work on other projects. However, it doesn’t really matter that much, since I am moving away from ReScript in favour of my planned second (and final, I promise!) rewrite in Clojure, so there will soon be a `clojure` branch in this repo.
The changes I have made mostly involve getting rid of the unnecessary distinction between expr<var> and expr<con> and therefore of the GADT type itself. I reflected about this in my Roam notebook (22.11.2021) and essentially it has the advantages of improving performance (the interpretation and the evaluation step can be merged → only one traversal), reducing cognitive load and simpler type-checking. The distinction can still be added later as an extension if needed. There is no real use-case or core feature that would justify this complexity and I made the error of premature abstraction here, even though it was kind-of fun to figure this stuff out.

If by some chance I (or someone else) continue(s) to work on this branch: there may be some unfinished parts that I have abandoned and some code from the var/con distinction might not have been completely wiped out. Feels bad to leave so much progress behind, but I am convinced that the new direction will be much more pragmatic, simple, flexible and interoperable without so much overhead on conceptual stuff.